refactor: move block building code to proposer.
