+++ lib/lambda_ethereum_consensus/validator/validator.ex
@@ -6,24 +6,19 @@ defmodule LambdaEthereumConsensus.Validator do
   require Logger
 
   alias LambdaEthereumConsensus.Beacon.BeaconChain
-  alias LambdaEthereumConsensus.Execution.ExecutionChain
-  alias LambdaEthereumConsensus.Execution.ExecutionClient
   alias LambdaEthereumConsensus.ForkChoice.Handlers
   alias LambdaEthereumConsensus.Libp2pPort
   alias LambdaEthereumConsensus.P2P.Gossip
   alias LambdaEthereumConsensus.StateTransition
   alias LambdaEthereumConsensus.StateTransition.Accessors
   alias LambdaEthereumConsensus.StateTransition.Misc
-  alias LambdaEthereumConsensus.Store.Blocks
   alias LambdaEthereumConsensus.Store.BlockStates
   alias LambdaEthereumConsensus.Utils.BitField
   alias LambdaEthereumConsensus.Utils.BitList
-  alias LambdaEthereumConsensus.Utils.Randao
   alias LambdaEthereumConsensus.Validator.BlockRequest
   alias LambdaEthereumConsensus.Validator.Proposer
   alias LambdaEthereumConsensus.Validator.Utils
   alias Types.Attestation
-  alias Types.BeaconState
 
   @default_graffiti_message "Lambda, so gentle, so good"
 
@@ -395,7 +390,7 @@ defmodule LambdaEthereumConsensus.Validator do
   end
 
   defp fetch_validator_index(beacon, %{index: nil, pubkey: pk}) do
-    beacon.validators |> Enum.find_index(&(&1.pubkey == pk))
+    Enum.find_index(beacon.validators, &(&1.pubkey == pk))
   end
 
   defp compute_proposer_duties(beacon_state, epoch, validator_index) do
@@ -412,79 +407,25 @@ defmodule LambdaEthereumConsensus.Validator do
   # Returns true if the validator should propose a block for the given slot
   defp should_propose?(%{duties: %{proposer: slots}}, slot), do: Enum.member?(slots, slot)
 
-  @spec build_execution_block(BeaconState.t(), Types.root()) ::
-          {:error, any()} | {:ok, Types.ExecutionPayload.t()}
-  def build_execution_block(state, head_root) do
-    head_block = Blocks.get_block!(head_root)
-    finalized_block = state.finalized_checkpoint.root |> Blocks.get_block!()
-
-    forkchoice_state = %{
-      finalized_block_hash: finalized_block.body.execution_payload.block_hash,
-      head_block_hash: head_block.body.execution_payload.block_hash,
-      # TODO calculate safe block
-      safe_block_hash: finalized_block.body.execution_payload.block_hash
-    }
-
-    payload_attributes = %{
-      timestamp: Misc.compute_timestamp_at_slot(state, state.slot),
-      prev_randao: Randao.get_randao_mix(state.randao_mixes, Accessors.get_current_epoch(state)),
-      suggested_fee_recipient: <<0::160>>,
-      # TODO: add withdrawals
-      withdrawals: [],
-      parent_beacon_block_root: head_root
-    }
-
-    with {:ok, payload_id} <-
-           ExecutionClient.notify_forkchoice_updated(
-             forkchoice_state,
-             payload_attributes
-           ),
-         # TODO: we need to balance a time that should be long enough to let the execution client
-         # pack as many transactions as possible (more fees for us) while giving enough time to propagate
-         # the block and have it included
-         :ok <- Process.sleep(3000),
-         {:ok, execution_payload} <-
-           ExecutionClient.get_payload(payload_id) do
-      {:ok, execution_payload}
-    end
-  end
-
   defp propose(%{root: head_root, validator: %{index: index, privkey: privkey}}, proposed_slot) do
-    head_state = BlockStates.get_state!(head_root)
+    # TODO: handle errors if there are any
+    {:ok, signed_block} =
+      Proposer.build_block(%BlockRequest{
+        slot: proposed_slot,
+        parent_root: head_root,
+        proposer_index: index,
+        graffiti_message: @default_graffiti_message,
+        privkey: privkey
+      })
 
-    with {:ok, execution_payload} <-
-           head_state |> process_slots(proposed_slot) |> build_execution_block(head_root) do
-      block_request =
-        %BlockRequest{
-          slot: proposed_slot,
-          parent_root: head_root,
-          proposer_index: index,
-          graffiti_message: @default_graffiti_message,
-          eth1_data: fetch_eth1_data(proposed_slot, head_state),
-          execution_payload: execution_payload
-        }
-        |> Map.merge(Proposer.fetch_operations_for_block())
+    {:ok, ssz_encoded} = Ssz.to_ssz(signed_block)
+    {:ok, encoded_msg} = :snappyer.compress(ssz_encoded)
+    fork_context = Base.encode16(BeaconChain.get_fork_digest(), case: :lower)
 
-      # TODO: handle errors if there are any
-      {:ok, signed_block} = Proposer.construct_block(head_state, block_request, privkey)
-
-      {:ok, ssz_encoded} = Ssz.to_ssz(signed_block)
-      {:ok, encoded_msg} = :snappyer.compress(ssz_encoded)
-      fork_context = BeaconChain.get_fork_digest() |> Base.encode16(case: :lower)
-
-      # TODO: we might want to send the block to ForkChoice
-
-      case Libp2pPort.publish("/eth2/#{fork_context}/beacon_block/ssz_snappy", encoded_msg) do
-        :ok -> Logger.info("[Validator] Proposed block for slot #{proposed_slot}")
-        _ -> Logger.error("[Validator] Failed to publish block for slot #{proposed_slot}")
-      end
-    end
-  end
-
-  defp fetch_eth1_data(slot, head_state) do
-    case ExecutionChain.get_eth1_vote(slot) do
-      nil -> head_state.eth1_data
-      eth1_data -> eth1_data
+    # TODO: we might want to send the block to ForkChoice
+    case Libp2pPort.publish("/eth2/#{fork_context}/beacon_block/ssz_snappy", encoded_msg) do
+      :ok -> Logger.info("[Validator] Proposed block for slot #{proposed_slot}")
+      _ -> Logger.error("[Validator] Failed to publish block for slot #{proposed_slot}")
     end
   end
 end
