+++ test/unit/validator/proposer_test.exs
@@ -40,7 +40,8 @@ defmodule Unit.Validator.ProposerTests do
         deposit_count: 64,
         block_hash: <<0::256>>
       },
-      execution_payload: spec_block.message.body.execution_payload
+      execution_payload: spec_block.message.body.execution_payload,
+      privkey: privkey
     }
 
     {:ok, signed_block} = Proposer.construct_block(pre_state, block_request, privkey)
